This PR introduces a comprehensive overhaul of the backend testing strategy, upgrading it to production-grade integration standards.

Key Changes :
Implemented Testcontainers-based integration testing for realistic database validation
Achieved 100% controller-level integration coverage
Eliminated ineffective/mock-heavy unit tests in favor of meaningful integration tests
Enforced strict validation, assertions, and type safety across the test suite
Fixed gaps in test lifecycle handling and security validation
Refactored overall testing architecture for improved scalability and maintainability
Added detailed documentation outlining the new testing strategy and best practices

Impact :
Significantly improves test reliability and accuracy
Ensures environment parity between tests and production
Reduces risk of false positives from shallow/mocked tests
Establishes a robust foundation for future backend development

+++ taskhive-backend/src/main/java/com/digiwork/taskhive/module/analytics/service/AnomalyDetectionService.java
@@ -63,7 +63,7 @@ public class AnomalyDetectionService {
             if (last7.isEmpty()) return;
 
             double avg7 = last7.stream()
-                    .mapToInt(m -> m.getCompletedTasks())
+                    .mapToDouble(m -> m.getCompletedTasks())
                     .average()
                     .orElse(0.0);
 
@@ -71,7 +71,7 @@ public class AnomalyDetectionService {
             var todayMetrics = dailyMetricsRepository.findTopByOrderByMetricDateDesc();
             if (todayMetrics.isEmpty()) return;
 
-            int todayCompleted = todayMetrics.get().getCompletedTasks();
+            long todayCompleted = todayMetrics.get().getCompletedTasks();
 
             if (avg7 > 0 && todayCompleted < avg7 * 0.6) {
                 activeAlerts.add(AnomalyAlertResponse.builder()
